requirements.txt
astroid==1.6.2
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.7.1
py==1.5.4
pylint==1.8.3
pytest==3.7.1
python-swiftclient==3.5.0
PyYAML==3.13
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
xar==18.7.12
